mbhd-brit/src/main/java/org/multibit/hd/brit/matcher/Matcher.java
package org.multibit.hd.brit.matcher;

import org.multibit.hd.brit.dto.EncryptedMatcherResponse;
import org.multibit.hd.brit.dto.EncryptedPayerRequest;
import org.multibit.hd.brit.dto.MatcherResponse;
import org.multibit.hd.brit.dto.PayerRequest;

/**
 *  <p>Interface to provide the following to BRIT:<br>
 *  <ul>
 *  <li>encapsulation of functionality required to match BRIT payers and redeemers</li>
 *  </ul>
 *  </p>
 *  
 */
public interface Matcher {

  public MatcherConfig getConfig();

  public PayerRequest decryptPayerRequest(EncryptedPayerRequest encryptedPayerRequest) throws Exception;

  public MatcherResponse process(PayerRequest payerRequest);

  public EncryptedMatcherResponse encryptMatcherResponse(MatcherResponse matcherResponse);
}
